editoast: move train schedules margins to schemas
